root/back-end/requirements.txt
Flask==1.0.3
https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp37-cp37m-linux_x86_64.whl
torchvision>=0.8.1
numpy>=1.18.5
Pillow==8.1.1
gunicorn==19.9.0
opencv-python-headless
pandas
requests
seaborn
flask-cors